Build with -Werror -Weverything, fix all warnings
Suppressed only warnings that don't apply to bare-metal GBA C:
  -Wno-padded, -Wno-unsafe-buffer-usage,
  -Wno-declaration-after-statement, -Wno-covered-switch-default,
  -Wno-switch-default, -Wno-unused-parameter

Fixed: sign conversions, implicit int narrowing, missing prototypes,
unused function/variable, unreachable returns, missing-variable-decl.
Use switch((int)x) where not all enum cases are handled.

+++ src/combat.c
@@ -53,7 +53,7 @@ void combat_update(void) {
 
     Entity *player = &entities[active_player];
 
-    switch (combat_state) {
+    switch ((int)combat_state) {
     case COMBAT_START:
         anim_timer++;
         if (anim_timer > 30) { // Brief "encounter!" flash
@@ -83,7 +83,7 @@ void combat_update(void) {
                 int dc = 2; // Base difficulty
                 CombatEnemy *target = &combat_enemies[target_cursor];
 
-                switch (selected_action) {
+                switch ((int)selected_action) {
                 case ACT_SHOOT:
                     stat = player->stats.shoot;
                     dc += target->stats.cool; // Enemy dodges with COOL
@@ -284,7 +284,7 @@ void combat_draw(void) {
     }
 
     // Draw based on state
-    switch (combat_state) {
+    switch ((int)combat_state) {
     case COMBAT_START:
         text_draw_string("!! ENCOUNTER !!", 24, 8, 9, 15);
         break;
